Rename JS files
- make activity.js the place for the activity logic
- make loader.js the place for the requirejs configuration
