+++ index.html
@@ -5,7 +5,7 @@
   <title>My Activity</title>
   <link rel="stylesheet" href="lib/sugar-html-graphics/css/sugar.css">
   <link rel="stylesheet" href="css/activity.css">
-  <script data-main="js/activity" src="lib/require.js"></script>
+  <script data-main="js/loader" src="lib/require.js"></script>
 </head>
 <body>
