+++ js/activity.js
@@ -1,8 +1,23 @@
-requirejs.config({
-    baseUrl: "lib",
-    paths: {
-        activity: "../js"
-    }
-});
+define(function (require) {
+    var activity = require("sugar-html-core/activity");
+    var icons = require("sugar-html-graphics/icons");
 
-requirejs(["activity/main"]);
+    // Initialize the activity.
+
+    activity.setup();
+
+    // Colorize the activity icon.
+
+    var activityButton = document.getElementById("activity-button");
+    activity.getXOColor(function (colors) {
+        icons.colorize(activityButton, colors);
+    });
+
+    // Make the activity stop with the stop button.
+
+    var stopButton = document.getElementById("stop-button");
+    stopButton.onclick = function () {
+        activity.close();
+    };
+
+});
