+++ js/loader.js
@@ -1,0 +1,8 @@
+requirejs.config({
+    baseUrl: "lib",
+    paths: {
+        activity: "../js"
+    }
+});
+
+requirejs(["activity/activity"]);
